Kernel/SysFS: Migrate components code from SysFS.cpp to the SysFS folder
