+++ Kernel/Devices/Device.h
@@ -22,7 +22,8 @@
 #include <Kernel/Devices/AsyncDeviceRequest.h>
 #include <Kernel/FileSystem/DeviceFileTypes.h>
 #include <Kernel/FileSystem/File.h>
-#include <Kernel/FileSystem/SysFS.h>
+#include <Kernel/FileSystem/SysFS/Registry.h>
+#include <Kernel/FileSystem/SysFS/Subsystems/Devices/DeviceComponent.h>
 #include <Kernel/UnixTypes.h>
 
 namespace Kernel {
