+++ Kernel/FileSystem/SysFS.cpp
@@ -4,69 +4,12 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <AK/Singleton.h>
 #include <AK/StringView.h>
-#include <Kernel/Devices/Device.h>
 #include <Kernel/FileSystem/SysFS.h>
+#include <Kernel/FileSystem/SysFS/Registry.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
-
-static Singleton<SysFSComponentRegistry> s_the;
-
-SysFSComponentRegistry& SysFSComponentRegistry::the()
-{
-    return *s_the;
-}
-
-UNMAP_AFTER_INIT void SysFSComponentRegistry::initialize()
-{
-    VERIFY(!s_the.is_initialized());
-    s_the.ensure_instance();
-}
-
-UNMAP_AFTER_INIT SysFSComponentRegistry::SysFSComponentRegistry()
-    : m_root_directory(SysFSRootDirectory::create())
-{
-}
-
-UNMAP_AFTER_INIT void SysFSComponentRegistry::register_new_component(SysFSComponent& component)
-{
-    MutexLocker locker(m_lock);
-    m_root_directory->m_components.append(component);
-}
-
-SysFSComponentRegistry::DevicesList& SysFSComponentRegistry::devices_list()
-{
-    return m_devices_list;
-}
-
-NonnullRefPtr<SysFSRootDirectory> SysFSRootDirectory::create()
-{
-    return adopt_ref(*new (nothrow) SysFSRootDirectory);
-}
-
-ErrorOr<void> SysFSRootDirectory::traverse_as_directory(FileSystemID fsid, Function<ErrorOr<void>(FileSystem::DirectoryEntryView const&)> callback) const
-{
-    MutexLocker locker(SysFSComponentRegistry::the().get_lock());
-    TRY(callback({ ".", { fsid, component_index() }, 0 }));
-    TRY(callback({ "..", { fsid, 0 }, 0 }));
-
-    for (auto const& component : m_components) {
-        InodeIdentifier identifier = { fsid, component.component_index() };
-        TRY(callback({ component.name(), identifier, 0 }));
-    }
-    return {};
-}
-
-SysFSRootDirectory::SysFSRootDirectory()
-{
-    auto buses_directory = SysFSBusDirectory::must_create(*this);
-    auto devices_directory = SysFSDevicesDirectory::must_create(*this);
-    m_components.append(buses_directory);
-    m_components.append(devices_directory);
-    m_buses_directory = buses_directory;
-}
 
 ErrorOr<NonnullRefPtr<FileSystem>> SysFS::try_create()
 {
@@ -224,28 +167,6 @@ ErrorOr<NonnullRefPtr<Inode>> SysFSDirectoryInode::lookup(StringView name)
     if (!component)
         return ENOENT;
     return TRY(component->to_inode(fs()));
-}
-
-SysFSBusDirectory& SysFSComponentRegistry::buses_directory()
-{
-    return *m_root_directory->m_buses_directory;
-}
-
-void SysFSComponentRegistry::register_new_bus_directory(SysFSDirectory& new_bus_directory)
-{
-    VERIFY(!m_root_directory->m_buses_directory.is_null());
-    m_root_directory->m_buses_directory->m_components.append(new_bus_directory);
-}
-
-UNMAP_AFTER_INIT NonnullRefPtr<SysFSBusDirectory> SysFSBusDirectory::must_create(SysFSRootDirectory const& parent_directory)
-{
-    auto directory = adopt_ref(*new (nothrow) SysFSBusDirectory(parent_directory));
-    return directory;
-}
-
-UNMAP_AFTER_INIT SysFSBusDirectory::SysFSBusDirectory(SysFSRootDirectory const& parent_directory)
-    : SysFSDirectory(parent_directory)
-{
 }
 
 }
