+++ Kernel/FileSystem/SysFS/RootDirectory.h
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2022, Liav A. <liavalb@hotmail.co.il>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <Kernel/FileSystem/SysFS/Component.h>
+#include <Kernel/FileSystem/SysFS/Subsystems/Bus/Directory.h>
+
+namespace Kernel {
+
+class SysFSRootDirectory final : public SysFSDirectory {
+    friend class SysFSComponentRegistry;
+
+public:
+    virtual StringView name() const override { return "."sv; }
+    static NonnullRefPtr<SysFSRootDirectory> create();
+    virtual ErrorOr<void> traverse_as_directory(FileSystemID, Function<ErrorOr<void>(FileSystem::DirectoryEntryView const&)>) const override;
+
+private:
+    SysFSRootDirectory();
+    RefPtr<SysFSBusDirectory> m_buses_directory;
+};
+
+}
