+++ Kernel/FileSystem/SysFS/Subsystems/Bus/PCI/BusDirectory.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include <Kernel/Bus/PCI/Definitions.h>
-#include <Kernel/FileSystem/SysFS.h>
+#include <Kernel/FileSystem/SysFS/Component.h>
 
 namespace Kernel {
